requirements.txt
customtkinter
paho-mqtt